requirements.txt
psycopg2-binary==2.7.5
Django==2.0.6
telepot==12.7
raven==6.9.0
PySocks==1.6.8
django-model-utils==3.1.2
pytz==2018.4
jsonfield==2.0.2
requests==2.18.4

celery==4.2.1
redis==2.10.6
django-redis-cache==1.7.1

django-post_office==3.0.4

django-cleanup==2.1.0

admin_logs==0.2.11
Logentries==0.17

pymessenger==0.0.7.0

urllib3==1.22

flower==0.9.2

# backup
django-dbbackup==3.2.0
boto3==1.7.47  # amazon library
django-storages==1.6.6

elastic-apm==3.0.1